+++ modules/core/src/test/java/org/gridgain/grid/spi/communication/tcp/GridTcpCommunicationSpiMultithreadedSelfTest.java
@@ -67,7 +67,7 @@ public abstract class GridTcpCommunicationSpiMultithreadedSelfTest extends GridS
 
     static {
         GridTcpCommunicationMessageFactory.registerCustom(new GridTcpCommunicationMessageProducer() {
-            @Override public GridTcpCommunicationMessageAdapter create(int type) {
+            @Override public GridTcpCommunicationMessageAdapter create(byte type) {
                 return new GridTestMessage();
             }
         }, GridTestMessage.DIRECT_TYPE);
